feat: 유저 리스트 컴포넌트 공통으로 분리
